training code updates

+++ system/forcefield.py
@@ -329,10 +329,11 @@ def parameterize(mol, forcefield):
     for k, v in vd.items():
         charge_param_idxs.append(v)
 
+#     scale_matrix = scale_matrix * 0
     nrg_fns.append((
         custom_ops.Electrostatics_f32,
         (
-            np.array(scale_matrix, dtype=np.int32),
+            np.array(scale_matrix, dtype=np.float32),
             np.array(charge_param_idxs, dtype=np.int32)
         )
     ))
